Added Climate zones & Mountains
Used the Abs function and set climate zones as an absolute number away from mid line of map. | Added Variance to this so that the cold regions bleed into the warmer ones. | Created a Mountain creation algorithm, but am  about to finesse this as, it is using lists and a random walker to create the mountain, it only creates single blobs of mountains and isn't very exciting / visually impressive.